requirements.txt
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
coloredlogs==15.0.1
Flask==2.3.1
flatbuffers==23.3.3
humanfriendly==10.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.3.0
numpy==1.24.3
onnxruntime==1.14.1
packaging==23.1
Pillow==9.5.0
protobuf==4.22.3
requests==2.29.0
sympy==1.11.1
urllib3==1.26.15
Werkzeug==2.3.0
